refactor: extract inline homepage sections into components/sections/
Extracted 7 sections and data utilities from pages/index.tsx (~680→~100 lines):

lib/repos.ts:
- Extracted getTopProjects, getVscodeExtensions, getNeovimPlugins, getCliTuiTools
- Exported Repo and RepoCategory types derived from repos.json

components/sections/:
- HeroSection.tsx - avatar, intro, social links
- AboutSection.tsx - about me with dynamic year calc
- ProjectsSection.tsx - top projects grid
- VscodeSection.tsx - VS Code extensions with ExtensionCard
- NeovimSection.tsx - Neovim plugins with NvimPluginCard
- CliTuiSection.tsx - CLI/TUI tools with CliToolCard
- ContactSection.tsx - contact/connect section

Each section is self-contained with null guards for empty data.
